Implement CRUD Services to support the Assignment entity Fixes #8
