+++ src/repository/person/person.go
@@ -1,0 +1,63 @@
+package repository
+
+import (
+	"github.com/gin-gonic/gin"
+
+	"madronetek.com/gradeit/config"
+	"madronetek.com/gradeit/model"
+)
+
+type repository struct{}
+
+// NewPerson is a factory function for store layer that returns a interface type, UserInt
+func NewPerson() PersonRepoInt {
+	return repository{}
+}
+
+// A RepositoryInt interface method
+//
+// Inserts a record in the user table
+func (repo repository) Create(ctx *gin.Context, person model.Person) {
+	result := config.DB.Create(&person)
+	if result.Error != nil {
+		ctx.JSON(500, gin.H{"error": result.Error})
+		return
+	}
+
+	ctx.JSON(200, gin.H{"data": person})
+}
+
+func (repo repository) GetAll(ctx *gin.Context) {
+	var people []model.Person
+
+	result := config.DB.Find(&people)
+	if result.Error != nil {
+		ctx.JSON(500, gin.H{"error": result.Error})
+	return
+	}
+
+	ctx.JSON(200, gin.H{"data": people})
+}
+
+func (repo repository) GetByID(ctx *gin.Context, id int64) {
+	var person model.Person
+
+	result := config.DB.First(&person, id)
+	if result.Error != nil {
+		ctx.JSON(500, gin.H{"error": result.Error})
+		return
+	}
+
+	ctx.JSON(200, gin.H{"data": person})
+}
+
+func (repo repository) Update(ctx *gin.Context, person model.Person) {
+	config.DB.Model(&person).Updates(model.Person{Name: person.Name, Email: person.Email, Password: person.Password})
+
+	ctx.JSON(200, gin.H{"data": person})
+}
+func (repo repository) Delete(ctx *gin.Context, id int64) {
+	config.DB.Delete(&model.Person{}, id)
+
+ 	ctx.JSON(200, gin.H{"data": "post has been deleted successfully"})
+}
